+++ internal/ast/css/core/CSSAtRule.ts
@@ -1,4 +1,4 @@
-import {CSSBlock, NodeBaseWithComments} from "@internal/ast";
+import {CSSBlock, CSSKeyframe, NodeBaseWithComments} from "@internal/ast";
 import {createBuilder} from "../../utils";
 import {AnyCSSValue} from "../../../css-parser/types";
 
@@ -6,7 +6,7 @@ export interface CSSAtRule extends NodeBaseWithComments {
 	readonly type: "CSSAtRule";
 	readonly name: string;
 	readonly prelude: AnyCSSValue[];
-	readonly block?: CSSBlock;
+	readonly block?: CSSBlock | CSSKeyframe;
 }
 export const cssAtRule = createBuilder<CSSAtRule>(
 	"CSSAtRule",
