+++ internal/css-parser/index.ts
@@ -1,6 +1,5 @@
 import {
 	AnyCSSToken,
-	AnyCSSValue,
 	CSSParser,
 	CSSParserOptions,
 	CSSParserTypes,
@@ -8,7 +7,6 @@ import {
 } from "./types";
 import {
 	TokenValues,
-	ValueToken,
 	createParser,
 	isDigit,
 	isHexDigit,
@@ -18,7 +16,6 @@ import {Number0, ob1Add, ob1Inc} from "@internal/ob1";
 import {
 	Symbols,
 	hexToUtf8,
-	isCustomProperty,
 	isIdentifierStart,
 	isName,
 	isNameStart,
@@ -27,19 +24,8 @@ import {
 	isValidEscape,
 	isWhitespace,
 } from "./utils";
-import {
-	CSSAtRule,
-	CSSBlock,
-	CSSCustomProperty,
-	CSSDeclaration,
-	CSSFunction,
-	CSSRoot,
-	CSSRule,
-	CSSSelector,
-	CSSVarFunction,
-} from "@internal/ast";
-import {parseSelectors} from "./parser/selectors";
-import {matchToken, nextToken, readToken} from "./tokenizer";
+import {CSSRoot} from "@internal/ast";
+import {parseRules} from "@internal/css-parser/parser/rules";
 
 export const cssParser = createParser<CSSParserTypes>({
 	diagnosticLanguage: "css",
@@ -634,543 +620,4 @@ export function parseCSS(opts: CSSParserOptions): CSSRoot {
 			body: rules,
 		}),
 	);
-}
-
-function parseRules(
-	parser: CSSParser,
-	topLevel = false,
-	endingTokenType?: keyof Tokens,
-): Array<CSSAtRule | CSSRule> {
-	const rules: Array<CSSAtRule | CSSRule> = [];
-	while (!matchToken(parser, "EOF")) {
-		if (endingTokenType && matchToken(parser, endingTokenType)) {
-			nextToken(parser);
-			break;
-		}
-
-		if (matchToken(parser, "Whitespace")) {
-			readToken(parser, "Whitespace");
-			continue;
-		}
-
-		if (matchToken(parser, "CDO") || matchToken(parser, "CDC")) {
-			if (topLevel) {
-				nextToken(parser);
-				continue;
-			}
-			const rule = parseRule(parser);
-			rule && rules.push(rule);
-			continue;
-		}
-
-		if (matchToken(parser, "AtKeyword")) {
-			rules.push(parseAtRule(parser));
-			continue;
-		}
-
-		const rule = parseRule(parser);
-		if (rule !== undefined) {
-			rules.push(rule);
-		}
-	}
-
-	return rules;
-}
-
-function parseRule(parser: CSSParser): CSSRule | undefined {
-	const start = parser.getPosition();
-	let prelude: CSSSelector[] = [];
-	while (!matchToken(parser, "EOF")) {
-		if (matchToken(parser, "LeftCurlyBracket")) {
-			return parser.finishNode(
-				start,
-				{
-					type: "CSSRule",
-					prelude,
-					block: parseDeclarationBlock(parser),
-				},
-			);
-		}
-		prelude = parseSelectors(parser);
-	}
-	parser.unexpectedDiagnostic({
-		description: descriptions.CSS_PARSER.UNEXPECTED_TOKEN,
-	});
-	return undefined;
-}
-
-function parseAtRule(parser: CSSParser): CSSAtRule {
-	const start = parser.getPosition();
-	const token = parser.expectToken("AtKeyword");
-	const prelude: AnyCSSValue[] = [];
-	const name = token.value;
-	let block = undefined;
-	while (true) {
-		if (matchToken(parser, "Semi")) {
-			break;
-		}
-		if (matchToken(parser, "EOF")) {
-			parser.unexpectedDiagnostic({
-				description: descriptions.CSS_PARSER.UNTERMINATED_AT_RULE,
-			});
-			break;
-		}
-		if (matchToken(parser, "LeftCurlyBracket")) {
-			block = parseComplexBlock(parser);
-			break;
-		}
-		const parsedValue = parseComponentValue(parser);
-		parsedValue && prelude.push(parsedValue);
-	}
-	return parser.finishNode(
-		start,
-		{
-			type: "CSSAtRule",
-			name,
-			prelude,
-			block,
-		},
-	);
-}
-
-function parseSimpleBlock(parser: CSSParser): CSSBlock | undefined {
-	const start = parser.getPosition();
-	const startingToken = parser.getToken();
-	const startingTokenValue = getBlockStartTokenValue(parser, startingToken);
-	const endingTokenType = getBlockEndTokenType(parser, startingToken);
-	let value: Array<AnyCSSValue | CSSAtRule | CSSDeclaration> = [];
-
-	if (!endingTokenType) {
-		return undefined;
-	}
-
-	nextToken(parser);
-
-	while (true) {
-		if (matchToken(parser, endingTokenType)) {
-			nextToken(parser);
-			break;
-		}
-		if (matchToken(parser, "EOF")) {
-			parser.unexpectedDiagnostic({
-				description: descriptions.CSS_PARSER.UNTERMINATED_BLOCK,
-			});
-			break;
-		}
-		const parsedValue = parseComponentValue(parser);
-		parsedValue && value.push(parsedValue);
-	}
-
-	return parser.finishNode(
-		start,
-		{
-			type: "CSSBlock",
-			startingTokenValue,
-			value,
-		},
-	);
-}
-
-function parseDeclarationBlock(parser: CSSParser): CSSBlock | undefined {
-	const start = parser.getPosition();
-	const startingToken = parser.getToken();
-	const startingTokenValue = getBlockStartTokenValue(parser, startingToken);
-	const endingTokenType = getBlockEndTokenType(parser, startingToken);
-	let value: Array<AnyCSSValue | CSSAtRule | CSSDeclaration> = [];
-
-	if (!endingTokenType) {
-		return undefined;
-	}
-
-	nextToken(parser);
-
-	value = parseDeclarations(parser, endingTokenType);
-
-	return parser.finishNode(
-		start,
-		{
-			type: "CSSBlock",
-			startingTokenValue,
-			value,
-		},
-	);
-}
-
-function parseComplexBlock(parser: CSSParser): CSSBlock | undefined {
-	const start = parser.getPosition();
-	const startingToken = parser.getToken();
-	const startingTokenValue = getBlockStartTokenValue(parser, startingToken);
-	const endingTokenType = getBlockEndTokenType(parser, startingToken);
-	let value: Array<CSSAtRule | CSSRule> = [];
-
-	if (!endingTokenType) {
-		return undefined;
-	}
-
-	nextToken(parser);
-
-	value = parseRules(parser, false, endingTokenType);
-
-	return parser.finishNode(
-		start,
-		{
-			type: "CSSBlock",
-			startingTokenValue,
-			value,
-		},
-	);
-}
-
-function parseComponentValue(parser: CSSParser): AnyCSSValue | undefined {
-	if (
-		matchToken(parser, "LeftCurlyBracket") ||
-		matchToken(parser, "LeftParen") ||
-		matchToken(parser, "LeftSquareBracket")
-	) {
-		return parseSimpleBlock(parser);
-	}
-
-	if (matchToken(parser, "Function")) {
-		return parseFunction(parser);
-	}
-
-	const start = parser.getPosition();
-
-	if (matchToken(parser, "Whitespace")) {
-		nextToken(parser);
-		return undefined;
-	}
-
-	if (matchToken(parser, "Dimension")) {
-		const unit = (parser.getToken() as Tokens["Dimension"]).unit;
-		const value = (parser.getToken() as Tokens["Dimension"]).value;
-		nextToken(parser);
-		return parser.finishNode(
-			start,
-			{
-				type: "CSSDimension",
-				unit,
-				value,
-			},
-		);
-	}
-
-	if (matchToken(parser, "Percentage")) {
-		const value = (parser.getToken() as Tokens["Percentage"]).value;
-		nextToken(parser);
-		return parser.finishNode(
-			start,
-			{
-				type: "CSSPercentage",
-				value,
-			},
-		);
-	}
-
-	if (matchToken(parser, "Ident")) {
-		const value = (parser.getToken() as Tokens["Ident"]).value;
-		nextToken(parser);
-		if (isCustomProperty(value)) {
-			return parser.finishNode(
-				parser.getPosition(),
-				{
-					type: "CSSCustomProperty",
-					value,
-				},
-			);
-		}
-
-		return parser.finishNode(
-			start,
-			{
-				type: "CSSIdentifier",
-				value,
-			},
-		);
-	}
-
-	if (matchToken(parser, "Number")) {
-		const numberToken = parser.getToken() as Tokens["Number"];
-		nextToken(parser);
-		return parser.finishNode(
-			start,
-			{
-				type: "CSSNumber",
-				value: numberToken.value,
-				raw: numberToken.raw,
-			},
-		);
-	}
-
-	if (matchToken(parser, "Colon")) {
-		nextToken(parser);
-		return parser.finishNode(
-			start,
-			{
-				type: "CSSRaw",
-				value: ":",
-			},
-		);
-	}
-
-	if (matchToken(parser, "Comma")) {
-		nextToken(parser);
-		return parser.finishNode(
-			start,
-			{
-				type: "CSSComma",
-			},
-		);
-	}
-
-	if (matchToken(parser, "Hash")) {
-		const hashToken = parser.getToken() as Tokens["Hash"];
-		if (hashToken.hashType === "id") {
-			nextToken(parser);
-			return parser.finishNode(
-				start,
-				{
-					type: "CSSHash",
-					value: `${hashToken.value}`,
-				},
-			);
-		}
-	}
-
-	if (matchToken(parser, "String")) {
-		const stringToken = parser.getToken() as Tokens["String"];
-		nextToken(parser);
-		return parser.finishNode(
-			start,
-			{
-				type: "CSSString",
-				value: stringToken.value,
-			},
-		);
-	}
-
-	const value = (parser.getToken() as ValueToken<string, string>).value;
-	nextToken(parser);
-	return parser.finishNode(
-		start,
-		{
-			type: "CSSRaw",
-			value,
-		},
-	);
-}
-
-export function parseFunction(parser: CSSParser): CSSFunction | CSSVarFunction {
-	const start = parser.getPosition();
-	const token = parser.getToken() as Tokens["Function"];
-	const name = token.value;
-	const params: AnyCSSValue[] = [];
-	const isVarFunction = name === "var";
-	parser.nextToken();
-
-	while (true) {
-		if (matchToken(parser, "RightParen")) {
-			nextToken(parser);
-			break;
-		}
-		if (matchToken(parser, "EOF")) {
-			parser.unexpectedDiagnostic({
-				description: descriptions.CSS_PARSER.UNTERMINATED_FUNCTION,
-			});
-			break;
-		}
-		const parsedValue = parseComponentValue(parser);
-		if (parsedValue) {
-			if (!params.length && isVarFunction) {
-				if (parsedValue.type !== "CSSCustomProperty") {
-					parser.unexpectedDiagnostic({
-						description: descriptions.CSS_PARSER.INVALID_CUSTOM_PROPERTY,
-					});
-				}
-				params.push(parsedValue);
-			} else {
-				params.push(parsedValue);
-			}
-		}
-	}
-
-	if (isVarFunction) {
-		return parser.finishNode(
-			start,
-			{
-				type: "CSSVarFunction",
-				name,
-				params: params as [CSSCustomProperty, ...AnyCSSValue[]],
-			},
-		);
-	}
-
-	return parser.finishNode(
-		start,
-		{
-			type: "CSSFunction",
-			name,
-			params,
-		},
-	);
-}
-
-function matchEndOfDeclaration(
-	parser: CSSParser,
-	endingTokenType: keyof Tokens,
-): boolean {
-	return (
-		matchToken(parser, "EOF") ||
-		matchToken(parser, "Semi") ||
-		matchToken(parser, endingTokenType)
-	);
-}
-
-function parseDeclarations(
-	parser: CSSParser,
-	endingTokenType: keyof Tokens,
-): Array<CSSAtRule | CSSDeclaration> {
-	const declarations: Array<CSSAtRule | CSSDeclaration> = [];
-
-	while (!matchToken(parser, "EOF")) {
-		if (readToken(parser, "Whitespace") || readToken(parser, "Semi")) {
-			continue;
-		}
-		if (matchToken(parser, endingTokenType)) {
-			nextToken(parser);
-			break;
-		}
-		if (matchToken(parser, "AtKeyword")) {
-			declarations.push(parseAtRule(parser));
-			continue;
-		}
-		if (matchToken(parser, "Ident")) {
-			while (!matchEndOfDeclaration(parser, endingTokenType)) {
-				const declaration = parseDeclaration(parser, endingTokenType);
-				declaration && declarations.push(declaration);
-			}
-			continue;
-		}
-		parser.unexpectedDiagnostic({
-			description: descriptions.CSS_PARSER.INVALID_DECLARATION,
-		});
-		while (!(matchToken(parser, "Semi") || matchToken(parser, "EOF"))) {
-			parseComponentValue(parser);
-		}
-	}
-
-	return declarations;
-}
-
-function parseDeclaration(
-	parser: CSSParser,
-	endingTokenType: keyof Tokens,
-): CSSDeclaration | undefined {
-	while (!matchEndOfDeclaration(parser, endingTokenType)) {
-		const currentToken = parser.getToken();
-		if (currentToken.type !== "Ident") {
-			parser.unexpectedDiagnostic({
-				description: descriptions.CSS_PARSER.INVALID_DECLARATION,
-			});
-			return undefined;
-		}
-		let name: string | CSSCustomProperty;
-		if (isCustomProperty(currentToken.value)) {
-			name = parser.finishNode(
-				parser.getPosition(),
-				{
-					type: "CSSCustomProperty",
-					value: currentToken.value,
-				},
-			);
-		} else {
-			name = currentToken.value;
-		}
-		const start = parser.getPosition();
-		let important = false;
-		let value: Array<AnyCSSValue | undefined> = [];
-		nextToken(parser);
-
-		while (matchToken(parser, "Whitespace")) {
-			readToken(parser, "Whitespace");
-		}
-		if (!matchToken(parser, "Colon")) {
-			parser.unexpectedDiagnostic({
-				description: descriptions.CSS_PARSER.INVALID_DECLARATION,
-			});
-			return undefined;
-		}
-		nextToken(parser);
-		while (matchToken(parser, "Whitespace")) {
-			readToken(parser, "Whitespace");
-		}
-		while (!matchEndOfDeclaration(parser, endingTokenType)) {
-			const parsedValue = parseComponentValue(parser);
-			parsedValue && value.push(parsedValue);
-		}
-
-		if (value.length >= 2) {
-			const lastTwoTokens = [...value].slice(-2);
-			if (
-				lastTwoTokens[0]?.type === "CSSRaw" &&
-				lastTwoTokens[0].value === "!" &&
-				lastTwoTokens[1]?.type === "CSSIdentifier" &&
-				/^important$/i.test(lastTwoTokens[1].value)
-			) {
-				value = value.slice(0, -2);
-				important = true;
-			}
-		}
-
-		return parser.finishNode(
-			start,
-			{
-				type: "CSSDeclaration",
-				important,
-				name,
-				value,
-			},
-		);
-	}
-	return undefined;
-}
-
-function getBlockStartTokenValue(
-	parser: CSSParser,
-	token: AnyCSSToken,
-): string | undefined {
-	switch (token.type) {
-		case "LeftCurlyBracket":
-			return "{";
-		case "LeftParen":
-			return "(";
-		case "LeftSquareBracket":
-			return "[";
-		default: {
-			parser.unexpectedDiagnostic({
-				description: descriptions.CSS_PARSER.INVALID_BLOCK_START,
-			});
-			return undefined;
-		}
-	}
-}
-
-function getBlockEndTokenType(
-	parser: CSSParser,
-	token: AnyCSSToken,
-): keyof Tokens | undefined {
-	switch (token.type) {
-		case "LeftCurlyBracket":
-			return "RightCurlyBracket";
-		case "LeftParen":
-			return "RightParen";
-		case "LeftSquareBracket":
-			return "RightSquareBracket";
-		default: {
-			parser.unexpectedDiagnostic({
-				description: descriptions.CSS_PARSER.INVALID_BLOCK_START,
-			});
-			return undefined;
-		}
-	}
 }
