+++ internal/css-parser/test-fixtures/invalid/keyfram-number/input.css
@@ -1,0 +1,5 @@
+@keyframes foo {
+	10 {
+
+	}
+}
