+++ internal/css-parser/test-fixtures/invalid/keyframe-css-wide-keywords/input.css
@@ -1,0 +1,3 @@
+@keyframes initial {}
+@keyframes inherit {}
+@keyframes unset {}
