+++ internal/css-parser/test-fixtures/invalid/keyframe-css-wide-keywords/input.test.md
@@ -1,0 +1,110 @@
+# `index.test.ts`
+
+**DO NOT MODIFY**. This file has been autogenerated. Run `rome test internal/css-parser/index.test.ts --update-snapshots` to update.
+
+## `invalid > keyframe-css-wide-keywords`
+
+```javascript
+CSSRoot {
+	comments: Array []
+	corrupt: false
+	filename: "invalid/keyframe-css-wide-keywords/input.css"
+	integrity: undefined
+	loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 1:0-3:19
+	body: Array [
+		CSSAtRule {
+			name: "keyframes"
+			block: undefined
+			prelude: Array []
+			loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 1:0-1:18
+		}
+		CSSRule {
+			prelude: Array []
+			loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 1:19-1:21
+			block: CSSBlock {
+				value: Array []
+				startingTokenValue: "{"
+				loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 1:19-1:21
+			}
+		}
+		CSSAtRule {
+			name: "keyframes"
+			block: undefined
+			prelude: Array []
+			loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 2:0-2:18
+		}
+		CSSRule {
+			prelude: Array []
+			loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 2:19-2:21
+			block: CSSBlock {
+				value: Array []
+				startingTokenValue: "{"
+				loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 2:19-2:21
+			}
+		}
+		CSSAtRule {
+			name: "keyframes"
+			block: undefined
+			prelude: Array []
+			loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 3:0-3:16
+		}
+		CSSRule {
+			prelude: Array []
+			loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 3:17-3:19
+			block: CSSBlock {
+				value: Array []
+				startingTokenValue: "{"
+				loc: SourceLocation invalid/keyframe-css-wide-keywords/input.css 3:17-3:19
+			}
+		}
+	]
+	diagnostics: Array [
+		Object {
+			origins: Array [Object {category: "parse"}]
+			location: Object {
+				filename: "invalid/keyframe-css-wide-keywords/input.css"
+				integrity: undefined
+				language: "css"
+				sourceText: undefined
+				end: Position 1:18
+				start: Position 1:11
+			}
+			description: Object {
+				category: "parse"
+				categoryValue: "css"
+				message: Array [
+					RAW_MARKUP {value: "The identifier <emphasis>"}
+					"initial"
+					RAW_MARKUP {value: "</emphasis> can't be used here"}
+				]
+				advice: Array [
+					log {
+						category: "info"
+						text: Array [
+							RAW_MARKUP {value: "In this position, the words "}
+							Array [
+								Array [
+									RAW_MARKUP {value: "<emphasis>"}
+									"unset"
+									RAW_MARKUP {value: "</emphasis>, "}
+								]
+								Array [
+									RAW_MARKUP {value: "<emphasis>"}
+									"initial"
+									RAW_MARKUP {value: "</emphasis>, "}
+								]
+								Array [
+									RAW_MARKUP {value: "<emphasis>"}
+									"inherit"
+									RAW_MARKUP {value: "</emphasis>, "}
+								]
+							]
+							RAW_MARKUP {value: " are CSS-wide keywords, so they are reserved"}
+						]
+					}
+				]
+			}
+		}
+	]
+}
+```
