+++ internal/css-parser/test-fixtures/invalid/keyframe-keyword/input.css
@@ -1,0 +1,7 @@
+@keyframes foo {
+	form {}
+}
+
+@keyframes foo {
+	ot {}
+}
