+++ internal/css-parser/test-fixtures/keyframe/input.test.md
@@ -1,0 +1,162 @@
+# `index.test.ts`
+
+**DO NOT MODIFY**. This file has been autogenerated. Run `rome test internal/css-parser/index.test.ts --update-snapshots` to update.
+
+## `keyframe`
+
+```javascript
+CSSRoot {
+	comments: Array []
+	corrupt: false
+	diagnostics: Array []
+	filename: "keyframe/input.css"
+	integrity: undefined
+	loc: SourceLocation keyframe/input.css 1:0-18:1
+	body: Array [
+		CSSAtRule {
+			name: "keyframes"
+			prelude: Array []
+			loc: SourceLocation keyframe/input.css 1:0-7:1
+			block: CSSKeyframe {
+				name: CSSKeyframeName {
+					value: CSSRaw {
+						value: "foo"
+						loc: SourceLocation keyframe/input.css 1:11-1:11
+					}
+					loc: SourceLocation keyframe/input.css 1:11-1:14
+				}
+				value: Array [
+					CSSKeyframeBlock {
+						name: CSSKeyframeSelector {
+							value: CSSRaw {
+								value: "from"
+								loc: SourceLocation keyframe/input.css 2:1-2:1
+							}
+							loc: SourceLocation keyframe/input.css 2:1-2:5
+						}
+						value: CSSBlock {
+							value: Array []
+							startingTokenValue: "{"
+							loc: SourceLocation keyframe/input.css 2:6-2:8
+						}
+						loc: SourceLocation keyframe/input.css 2:8-2:8
+					}
+					CSSKeyframeBlock {
+						name: CSSKeyframeSelector {
+							value: CSSRaw {
+								value: "to"
+								loc: SourceLocation keyframe/input.css 3:1-3:1
+							}
+							loc: SourceLocation keyframe/input.css 3:1-3:3
+						}
+						value: CSSBlock {
+							value: Array []
+							startingTokenValue: "{"
+							loc: SourceLocation keyframe/input.css 3:4-3:6
+						}
+						loc: SourceLocation keyframe/input.css 3:6-3:6
+					}
+					CSSKeyframeBlock {
+						name: CSSKeyframeSelector {
+							value: CSSPercentage {
+								value: 15
+								loc: SourceLocation keyframe/input.css 4:1-4:1
+							}
+							loc: SourceLocation keyframe/input.css 4:1-4:4
+						}
+						value: CSSBlock {
+							value: Array []
+							startingTokenValue: "{"
+							loc: SourceLocation keyframe/input.css 4:5-4:7
+						}
+						loc: SourceLocation keyframe/input.css 4:7-4:7
+					}
+					CSSKeyframeBlock {
+						name: CSSKeyframeSelector {
+							value: CSSPercentage {
+								value: 0
+								loc: SourceLocation keyframe/input.css 5:1-5:1
+							}
+							loc: SourceLocation keyframe/input.css 5:1-5:3
+						}
+						value: CSSBlock {
+							value: Array []
+							startingTokenValue: "{"
+							loc: SourceLocation keyframe/input.css 5:4-5:6
+						}
+						loc: SourceLocation keyframe/input.css 5:6-5:6
+					}
+					CSSKeyframeBlock {
+						name: CSSKeyframeSelector {
+							value: CSSPercentage {
+								value: 100
+								loc: SourceLocation keyframe/input.css 6:1-6:1
+							}
+							loc: SourceLocation keyframe/input.css 6:1-6:5
+						}
+						value: CSSBlock {
+							value: Array []
+							startingTokenValue: "{"
+							loc: SourceLocation keyframe/input.css 6:6-6:8
+						}
+						loc: SourceLocation keyframe/input.css 6:8-6:8
+					}
+				]
+				startingTokenValue: "{"
+				loc: SourceLocation keyframe/input.css 1:10-7:1
+			}
+		}
+		CSSAtRule {
+			name: "keyframes"
+			prelude: Array []
+			loc: SourceLocation keyframe/input.css 9:0-11:1
+			block: CSSKeyframe {
+				name: CSSKeyframeName {
+					value: CSSString {
+						value: "foo"
+						loc: SourceLocation keyframe/input.css 9:11-9:11
+					}
+					loc: SourceLocation keyframe/input.css 9:11-9:16
+				}
+				value: Array []
+				startingTokenValue: "{"
+				loc: SourceLocation keyframe/input.css 9:10-11:1
+			}
+		}
+		CSSAtRule {
+			name: "keyframes"
+			prelude: Array []
+			loc: SourceLocation keyframe/input.css 13:0-15:1
+			block: CSSKeyframe {
+				name: CSSKeyframeName {
+					value: CSSString {
+						value: "Initial"
+						loc: SourceLocation keyframe/input.css 13:11-13:11
+					}
+					loc: SourceLocation keyframe/input.css 13:11-13:20
+				}
+				value: Array []
+				startingTokenValue: "{"
+				loc: SourceLocation keyframe/input.css 13:10-15:1
+			}
+		}
+		CSSAtRule {
+			name: "keyframes"
+			prelude: Array []
+			loc: SourceLocation keyframe/input.css 17:0-18:1
+			block: CSSKeyframe {
+				name: CSSKeyframeName {
+					value: CSSRaw {
+						value: "FOO"
+						loc: SourceLocation keyframe/input.css 17:11-17:11
+					}
+					loc: SourceLocation keyframe/input.css 17:11-17:14
+				}
+				value: Array []
+				startingTokenValue: "{"
+				loc: SourceLocation keyframe/input.css 17:10-18:1
+			}
+		}
+	]
+}
+```
