+++ internal/formatter/builders/css/core/CSSAtRule.ts
@@ -1,6 +1,11 @@
 import {CSSAtRule} from "@internal/ast";
-import {Builder, Token} from "@internal/formatter";
+import {Builder, Token, concat, space} from "@internal/formatter";
 
 export default function CSSAtRule(builder: Builder, node: CSSAtRule): Token {
-	throw new Error("unimplemented");
+	return concat([
+		...node.prelude.map((token) => builder.tokenize(token, node)),
+
+		space,
+		builder.tokenize(node.block, node),
+	]);
 }
