+++ internal/formatter/builders/css/keyframe/CSSKeyframeName.ts
@@ -1,0 +1,9 @@
+import {CSSKeyframeName} from "@internal/ast";
+import {Builder, Token} from "@internal/formatter";
+
+export default function CSSKeyframeName(
+	builder: Builder,
+	node: CSSKeyframeName,
+): Token {
+	return builder.tokenize(node.value, node);
+}
