+++ internal/formatter/test-fixtures/css/keyframe/input.css
@@ -1,0 +1,14 @@
+@keyframes    foo
+
+{
+  from    {  opacity: 0; }
+
+	15%
+
+	{
+		opacity: 0.23;
+	}
+
+	to {
+		background-color: red;}
+ }
