+++ internal/formatter/test-fixtures/css/keyframe/input.test.md
@@ -1,0 +1,51 @@
+# `index.test.ts`
+
+**DO NOT MODIFY**. This file has been autogenerated. Run `rome test internal/formatter/index.test.ts --update-snapshots` to update.
+
+## `css > keyframe`
+
+### `Diagnostics`
+
+```
+✔ No known problems!
+
+```
+
+### `Input`
+
+```css
+@keyframes    foo
+
+{
+  from    {  opacity: 0; }
+
+	15%
+
+	{
+		opacity: 0.23;
+	}
+
+	to {
+		background-color: red;}
+ }
+
+```
+
+### `Output`
+
+```css
+ @keyframes foo {
+	from {
+		opacity: 0;
+	}
+
+	15% {
+		opacity: 0.23;
+	}
+
+	to {
+		background-color: red;
+	}
+}
+
+```
